feat(engine): direct per-content palette/atlas handles — remove palette & atlas sets
Each TileCell and Sprite names its own sheet (AtlasId) and palette (PaletteId)
directly as 16-bit handles. The per-layer palette set and atlas set, the
per-cell/per-sprite selects, and the 16-sheet/16-palette cap are removed. The
tilemap cell is a two-word R32G32_UINT record; GpuSprite packs the atlas and
palette handles into one word; the per-layer set uniforms are replaced by a
global atlas-region table the tile and sprite shaders index by a cell's/sprite's
atlas handle. assembleTilemap emits cells with direct handles; a tiles() helper
fills the repeated handles over a single-combo run. Sprites name their atlas
per-sprite, so one sprite layer mixes sheets and palettes freely.

Behaviour-preserving: the golden GPU readback reproduces prior output on Metal,
Vulkan, and D3D12. The renderer, the tile/sprite shaders, all unit tests, the
demos, and the guide docs are reconciled to the new model. CI green on all 5 jobs.

+++ include/retropp/image.h
@@ -90,12 +90,14 @@ struct IndexGrid {
 // convenience that chains loadPng → uploadAtlas → sliceLayout into an AtlasManifest.
 
 // A handle to uploaded atlas pixel data the renderer owns. Identity is the typed handle; the
-// renderer maps it to its GPU texture. Sibling to PaletteId (palette.h). It lives here in image.h —
-// beside the atlas-ingestion surface (AssetSlot / AtlasManifest), the semantically correct home for
-// a handle to uploaded atlas *image* data — so a consumer can name a frame's atlas (animation.h)
-// without pulling in the whole draw-state submission envelope. draw_state.h includes image.h to use
-// it (TileContent / SpriteContent carry an AtlasId); the fully-qualified name is unchanged.
-enum class AtlasId : std::uint32_t {};
+// renderer maps it (via the global atlas-region table) to its placement in the flat atlas store.
+// Sibling to PaletteId (palette.h). It lives here in image.h — beside the atlas-ingestion surface
+// (AssetSlot / AtlasManifest), the semantically correct home for a handle to uploaded atlas *image*
+// data — so a consumer can name a frame's atlas (animation.h) without pulling in the whole draw-state
+// submission envelope. draw_state.h includes image.h to use it (TileCell / Sprite carry an AtlasId);
+// the fully-qualified name is unchanged. 16-bit: a tilemap cell carries it directly every frame, so it
+// is sized to the real ceiling (thousands of sheets) with headroom, not to 32-bit billions.
+enum class AtlasId : std::uint16_t {};
 
 // The atlas addressing cell: 8px. This is the atomic tile/OBJ cell of the whole 8/16-bit era — GB/GBC,
 // NES, SMS, SNES, and Genesis all dice their art into 8×8 cells, and nothing in the paradigm is finer —
